Fix kernel_dot_part2 for smaller BLOCKSIZE (#87)
* Switch build to c++17

* Update clang-format to use c++17

* Fix OOB LDS access for larger BLOCKSIZE

This kernel is called with both BLOCKSIZE 1024 and 256. For the
256 case, the first two if-statements are trivially true, but then
we access entries in LDS outside of the allocated LDS space. We
leverage the constexpr-if idiom in C++17 to determine at compile
time if we need the additional checks at larger BLOCKSIZE.

* Ensure BLOCKSIZE is adhering to the assumption

So hopefullly we don't make the same mistake again.